requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Werkzeug==3.0.1
Pillow==10.2.0
python-socketio==5.11.0
eventlet==0.35.1
httpx==0.27.0
